api/requirements.txt
fastapi==0.95.1
logger==1.4
neo4j==5.8.0
uvicorn==0.22.0
openai
google-generativeai
retry==0.9.2
tiktoken==0.4.0
python-dotenv==1.0.0
websockets===11.0.3
gunicorn===20.1.0